requirements.txt
matplotlib==3.10.1
numpy==2.2.4
Pillow==11.1.0
ptflops==0.7.4
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.6.1
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
transformers==4.46.1
